refactor: remove manual disk space error annotation in favor of direct OS error detection via internal/utils
