--- internal/strategy/concurrent/enospc_test_unix.go
@@ -1,12 +1,0 @@
-//go:build unix
-
-package concurrent
-
-import (
-	"os"
-	"syscall"
-)
-
-func makeDiskFullPathError() error {
-	return &os.PathError{Op: "write", Path: "x", Err: syscall.ENOSPC}
-}
