--- internal/strategy/single/enospc_annotate_test_unix.go
@@ -1,12 +1,0 @@
-//go:build unix
-
-package single
-
-import (
-	"os"
-	"syscall"
-)
-
-func makeDiskFullPathError() error {
-	return &os.PathError{Op: "write", Path: "x", Err: syscall.ENOSPC}
-}
