--- internal/strategy/single/enospc_annotate_test_windows.go
@@ -1,13 +1,0 @@
-//go:build windows
-
-package single
-
-import (
-	"os"
-
-	"golang.org/x/sys/windows"
-)
-
-func makeDiskFullPathError() error {
-	return &os.PathError{Op: "write", Path: "x", Err: windows.ERROR_DISK_FULL}
-}
