--- internal/types/disk_space_test_unix.go
@@ -1,12 +1,0 @@
-//go:build unix
-
-package types
-
-import (
-	"os"
-	"syscall"
-)
-
-func makeDiskFullPathError() error {
-	return &os.PathError{Op: "write", Path: "x", Err: syscall.ENOSPC}
-}
