--- internal/types/disk_space_windows.go
@@ -1,18 +1,0 @@
-//go:build windows
-
-package types
-
-import (
-	"errors"
-	"syscall"
-
-	"golang.org/x/sys/windows"
-)
-
-func isDiskFull(err error) bool {
-	var errno syscall.Errno
-	if !errors.As(err, &errno) {
-		return false
-	}
-	return errno == windows.ERROR_DISK_FULL || errno == windows.ERROR_DISK_QUOTA_EXCEEDED
-}
