+++ internal/utils/disk_free_unix.go
@@ -2,7 +2,19 @@
 
 package utils
 
-import "syscall"
+import (
+	"errors"
+	"syscall"
+)
+
+// IsOSDiskFull reports whether the error is a unix disk full error.
+func IsOSDiskFull(err error) bool {
+	var errno syscall.Errno
+	if !errors.As(err, &errno) {
+		return false
+	}
+	return errno == syscall.ENOSPC || errno == syscall.EDQUOT
+}
 
 func freeDiskBytesAt(path string) (int64, error) {
 	var st syscall.Statfs_t
